separate mantis directory notebook
