+++ python_files/create_fov_stats.py
@@ -315,180 +315,4 @@ keep_dict = {'CD38': lymphocyte + monocyte + stroma + granulocyte, 'CD45RB': lym
              'PD1': lymphocyte, 'PDL1': lymphocyte + monocyte + granulocyte + cancer, 'PDL1_tumor_dim': cancer,
              'TBET': lymphocyte, 'TCF1': lymphocyte, 'TIM3': lymphocyte + monocyte + granulocyte}
 
-# create mantis_dir to inspect invididual FOVs
-func_df = pd.read_csv(os.path.join(data_dir, 'combined_cell_table_normalized_cell_labels_updated_functional_only.csv'))
-
-data_wide.loc[data_wide['GLUT1+_Cancer'] > 0.25, ['GLUT1+_M2_Mac', 'GLUT1+_Fibroblast', 'GLUT1+_Cancer']]
-fovs = data_wide.loc[data_wide['HLA1+_Treg'] > 0.5, :].index.tolist()
-
-# select 5th and 9th fov
-fovs = fovs[0:2] + fovs[5:7] + fovs[9:11]
-fovs.remove('TONIC_TMA23_R10C1')
-
-fovs = ['TONIC_TMA9_R6C3', 'TONIC_TMA9_R7C3', 'TONIC_TMA10_R11C6', 'TONIC_TMA10_R5C5', 'TONIC_TMA16_R3C4', 'TONIC_TMA17_R12C3', 'TONIC_TMA3_R2C4']
-fovs = []
-
-marker = 'GLUT1'
-cell_types = ['Cancer', 'Fibroblast', 'Cancer_EMT', 'Cancer_Other', 'M2_Mac']
-input_cell_table = func_df.copy()
-input_cell_table['cell_cluster_new'] = np.where(func_df.cell_cluster.isin(cell_types), func_df.cell_cluster, 'Other')
-input_cell_table['cell_cluster_new'] = input_cell_table['cell_cluster_new'].values + input_cell_table[marker].astype('str')
-input_cell_table['cell_cluster_new'] = np.where(input_cell_table['PDL1_tumor_dim'].values, input_cell_table['cell_cluster'] + input_cell_table['PDL1_tumor_dim'].astype('str') + '_dim', input_cell_table['cell_cluster_new'])
-
-keep_channels = ['CD38.tiff']
-create_mantis_project(cell_table=input_cell_table, fovs=fovs, seg_dir='/Volumes/Shared/Noah Greenwald/TONIC_Cohort/segmentation_data/deepcell_output',
-                      pop_col='cell_cluster_new', mask_dir='/Volumes/Shared/Noah Greenwald/TONIC_Cohort/mantis_dir/masks',
-                      image_dir='/Volumes/Shared/Noah Greenwald/TONIC_Cohort/image_data/samples', mantis_dir='/Volumes/Shared/Noah Greenwald/TONIC_Cohort/mantis_dir/mantis_folders')
-all_chans = io_utils.list_files('/Volumes/Shared/Noah Greenwald/TONIC_Cohort/image_data/samples/TONIC_TMA3_R4C2', '.tiff')
-# TONIC_TMA2_R8C2
-for chan in all_chans:
-    if chan not in keep_channels:
-        for fov in fovs:
-            path = os.path.join('/Volumes/Shared/Noah Greenwald/TONIC_Cohort/mantis_dir/mantis_folders', fov, chan)
-            if os.path.exists(path):
-                os.remove(path)
-
-from ark.utils import data_utils, plot_utils, load_utils, io_utils
-from ark.utils.misc_utils import verify_in_list
-from ark import settings
-
-import skimage.io as io
-
-
-def label_cells_by_cluster(fov, all_data, label_map, fov_col=settings.FOV_ID,
-                           cell_label_column=settings.CELL_LABEL,
-                           cluster_column=settings.KMEANS_CLUSTER):
-    """Translates cell-ID labeled images according to the clustering assignment.
-    Takes a single FOV, and relabels the image according to the assignment
-    of cell IDs to cluster label.
-    Args:
-        fov (str):
-            The FOV to relabel
-        all_data (pandas.DataFrame):
-            data including fovs, cell labels, and cell expression matrix for all markers.
-        label_map (xarray.DataArray):
-            label map for a single FOV
-        fov_col (str):
-            column with the fovs names in `all_data`.
-        cell_label_column (str):
-            column with the cell labels in `all_data`.
-        cluster_column (str):
-            column with the cluster labels in `all_data`.
-    Returns:
-        numpy.ndarray:
-            The image with new designated label assignments
-    """
-
-    # verify that fov found in all_data
-    # NOTE: label_map fov validation happens in loading function
-    verify_in_list(fov_name=[fov], all_data_fovs=all_data[fov_col].unique())
-
-    # subset all_data on the FOV
-    df = all_data[all_data[fov_col] == fov]
-
-    # generate the labels to use
-    labels_dict = dict(zip(df[cell_label_column], df[cluster_column]))
-
-    # condense extraneous axes
-    labeled_img_array = label_map.squeeze().values
-
-    # relabel the array
-    relabeled_img_array = data_utils.relabel_segmentation(labeled_img_array, labels_dict)
-
-    return relabeled_img_array
-
-
-def save_fov_mask(fov, data_dir, mask_data, sub_dir=None, name_suffix=''):
-    """Saves a provided cluster label mask overlay for a FOV.
-    Args:
-        fov (str):
-            The FOV to save
-        data_dir (str):
-            The directory to save the cluster mask
-        mask_data (numpy.ndarray):
-            The cluster mask data for the FOV
-        sub_dir (Optional[str]):
-            The subdirectory to save the masks in. If specified images are saved to
-            "data_dir/sub_dir". If `sub_dir = None` the images are saved to `"data_dir"`.
-            Defaults to `None`.
-        name_suffix (str):
-            Specify what to append at the end of every fov.
-    """
-
-    # data_dir validation
-    io_utils.validate_paths(data_dir)
-
-    # ensure None is handled correctly in file path generation
-    if sub_dir is None:
-        sub_dir = ''
-
-    save_dir = os.path.join(data_dir, sub_dir)
-
-    # make the save_dir if it doesn't already exist
-    if not os.path.exists(save_dir):
-        os.makedirs(save_dir)
-
-    # define the file name as the fov name with the name suffix appended
-    fov_file = fov + name_suffix + '.tiff'
-
-    # save the image to data_dir
-    io.imsave(os.path.join(save_dir, fov_file), mask_data, check_contrast=False)
-
-def create_mantis_project(cell_table, fovs, seg_dir, pop_col,
-                          mask_dir, image_dir, mantis_dir) -> None:
-    """Create a complete Mantis project for viewing cell labels
-    Args:
-        cell_table (pd.DataFrame): dataframe of extracted cell features and subtypes
-        fovs (list): list of FOVs to use for creating the project
-        seg_dir (path): path to the directory containing the segmentations
-        pop_col (str): the column containing the distinct cell populations
-        mask_dir (path): path to the directory where the masks will be stored
-        image_dir (path): path to the directory containing the raw image data
-        mantis_dir (path): path to the directory where the mantis project will be created
-        seg_suffix_name (str, optional):
-            The suffix of the segmentation file. Defaults to "_whole_cell".
-    """
-
-    if not os.path.exists(mask_dir):
-        os.makedirs(mask_dir)
-
-    # create small df compatible with FOV function
-    small_table = cell_table.loc[:, [pop_col, 'label', 'fov']]
-
-    # generate unique numeric value for each population
-    small_table['pop_vals'] = pd.factorize(small_table[pop_col].tolist())[0] + 1
-
-    # label and save the cell mask for each FOV
-    for fov in fovs:
-        whole_cell_file = [fov + '_feature_0.tiff' for fov in fovs]
-        print(whole_cell_file)
-        # load the segmentation labels in for the FOV
-        label_map = load_utils.load_imgs_from_dir(
-            data_dir=seg_dir, files=whole_cell_file, xr_dim_name='compartments',
-            xr_channel_names=['feature_0'], trim_suffix='_feature_0'
-        ).loc[fov, ...]
-
-        # use label_cells_by_cluster to create cell masks
-        mask_data = label_cells_by_cluster(
-            fov, small_table, label_map, fov_col='fov',
-            cell_label_column='label', cluster_column='pop_vals'
-        )
-
-        # save the cell mask for each FOV
-        save_fov_mask(
-            fov,
-            mask_dir,
-            mask_data,
-            sub_dir=None,
-            name_suffix='_cell_mask'
-        )
-
-    # rename the columns of small_table
-    mantis_df = small_table.rename({'pop_vals': 'metacluster', pop_col: 'mc_name'}, axis=1)
-
-    # create the mantis project
-    plot_utils.create_mantis_dir(fovs=fovs, mantis_project_path=mantis_dir,
-                                 img_data_path=image_dir, mask_output_dir=mask_dir,
-                                 mask_suffix='_cell_mask', mapping=mantis_df,
-                                 seg_dir=seg_dir, img_sub_folder='')
